+++ Threeyes/SDK/Scripts/Component/BuiltIn/Animation/SO/AC_SOCharacterAnimatorControllerConfig.cs
@@ -1,0 +1,6 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "Animation/CharacterAnimatorController", fileName = "CharacterAnimatorControllerConfig")]
+public class AC_SOCharacterAnimatorControllerConfig : AC_SOConfigBase<AC_CharacterAnimatorController.ConfigInfo> { }