+++ Threeyes/SDK/Scripts/Component/BuiltIn/Transform/IAC_ObjectMovement.cs
@@ -1,0 +1,20 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+/// <summary>
+/// Control object movement
+/// </summary>
+public interface IAC_ObjectMovement
+{
+	/// <summary>
+	/// Check if the target moving 
+	/// </summary>
+	bool IsMoving { get; }
+
+	
+	float CurMoveSpeedPercent { get; }
+	float CurMoveSpeed { get; }
+	float MaxMoveSpeed { get; }
+	float LastMoveTime { get; }
+}