+++ Threeyes/SDK/Scripts/Component/BuiltIn/Transform/SO/AC_SOObjectMovement_FollowTargetConfig.cs
@@ -1,0 +1,6 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config + "Transform/ObjectMovement/FollowTarget", fileName = "ObjectMovement_FollowTargetConfig")]
+public class AC_SOObjectMovement_FollowTargetConfig : AC_SOConfigBase<AC_ObjectMovement_FollowTarget.ConfigInfo> { }